move about-us elements together
